+++ src/SeedsOfChange.cpp
@@ -47,8 +47,17 @@ struct SeedsOfChange : Module {
 		NUM_LIGHTS
 	};
 
+
 	float outbuffer[NBOUT * 2];
-	
+
+
+	// Expander
+	float consumerMessage[4] = {};// this module must read from here
+	float producerMessage[4] = {};// mother will write into here
+	int seed;
+	float clockValue,resetValue,distributionValue;
+
+
 	dsp::SchmittTrigger resetTrigger,clockTrigger,distributionModeTrigger; 
 
 	bool gaussianMode = false;
@@ -64,6 +73,10 @@ struct SeedsOfChange : Module {
 			configParam(SeedsOfChange::OFFSET_1_PARAM + i, -10.0f, 10.0f, 0.0f,"Offset");			
 			configParam(SeedsOfChange::GATE_PROBABILITY_1_PARAM + i, 0.0, 1.0, 0.0,"Gate Probability","%",0,100);
 		}
+
+		rightExpander.producerMessage = producerMessage;
+		rightExpander.consumerMessage = consumerMessage;
+
 	}
 	unsigned long mt[N]; /* the array for the state vector  */
 	int mti=N+1; /* mti==N+1 means mt[N] is not initialized */
@@ -127,7 +140,20 @@ struct SeedsOfChange : Module {
 		for (int i=0; i<NBOUT; i++) {
 			outputs[CV_1_OUTPUT+i].value = outbuffer[i];
 			outputs[GATE_1_OUTPUT+i].value = outbuffer[i+NBOUT] ? inputs[CLOCK_INPUT].value : 0;
-		}								
+		}	
+
+		//Set Expander Info
+		if(rightExpander.module && rightExpander.module->model == modelSeedsOfChangeCVExpander) {	
+
+			//Send outputs to slaves if present		
+			float *messageToExpander = (float*)(rightExpander.module->leftExpander.producerMessage);
+			
+			messageToExpander[0] = latest_seed; 
+			messageToExpander[1] = inputs[CLOCK_INPUT].getVoltage(); 
+			messageToExpander[2] = resetInput; 
+			messageToExpander[3] = gaussianMode; 
+			rightExpander.module->leftExpander.messageFlipRequested = true;						
+		}					
 	}
 
 	// For more advanced Module features, see engine/Module.hpp in the Rack API.